backend/requirements.txt
fastapi
uvicorn
transformers
torch
scispacy
spacy
pydantic
python-dotenv
crewai
langchainpydantic-settings
